+++ CHANGELOG.md
@@ -6,6 +6,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.10.0] - 2023-01-07
+
+### Added
+
+- Map are now created with a random algorithm
+- Cellular Automata algorithm to create more organic map
+- Drunkard algorithm to create cave like map
+- Empty map to create an empty room
+- Prefab rooms to inject into map
+- Theme for map (Forest, Dungeon)
+- Chose random Theme for generated map
+
+### Changed
+
+- MapBuilder now handles multi map generator algorithms 
+
 ## [0.9.0] - 2023-01-05
 
 ### Added
@@ -110,7 +126,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Camera which follow the Player
 - Graphics to all Tiles
 
-[unreleased]: https://github.com/ulphidius/dungeoncrawl/compare/v0.9.0...master
+[unreleased]: https://github.com/ulphidius/dungeoncrawl/compare/v0.10.0...master
+[0.10.0]: https://github.com/ulphidius/dungeoncrawl/compare/v0.9.0...v0.10.0
 [0.9.0]: https://github.com/ulphidius/dungeoncrawl/compare/v0.8.1...v0.9.0
 [0.8.1]: https://github.com/ulphidius/dungeoncrawl/compare/v0.8.0...v0.8.1
 [0.8.0]: https://github.com/ulphidius/dungeoncrawl/compare/v0.7.0...v0.8.0